+++ src/audiodevice.rs
@@ -37,7 +37,7 @@ use crate::CommandMessage;
 use crate::PrcFmt;
 use crate::Res;
 use crate::StatusMessage;
-use crate::{CaptureStatus, PlaybackStatus};
+use crate::{CaptureStatus, PlaybackStatus, ProcessingParameters};
 
 pub const RATE_CHANGE_THRESHOLD_COUNT: usize = 3;
 pub const RATE_CHANGE_THRESHOLD_VALUE: f32 = 0.04;
@@ -238,6 +238,7 @@ pub trait CaptureDevice {
         status_channel: crossbeam_channel::Sender<StatusMessage>,
         command_channel: mpsc::Receiver<CommandMessage>,
         capture_status: Arc<RwLock<CaptureStatus>>,
+        processing_params: Arc<ProcessingParameters>,
     ) -> Res<Box<thread::JoinHandle<()>>>;
 }
 
@@ -555,7 +556,9 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             ref device,
             format,
             stop_on_inactive,
-            ref follow_volume_control,
+            ref link_volume_control,
+            ref link_mute_control,
+            ..
         } => Box::new(alsadevice::AlsaCaptureDevice {
             devname: device.clone(),
             samplerate: conf.samplerate,
@@ -569,13 +572,15 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             stop_on_rate_change: conf.stop_on_rate_change(),
             rate_measure_interval: conf.rate_measure_interval(),
             stop_on_inactive: stop_on_inactive.unwrap_or_default(),
-            follow_volume_control: follow_volume_control.clone(),
+            link_volume_control: link_volume_control.clone(),
+            link_mute_control: link_mute_control.clone(),
         }),
         #[cfg(feature = "pulse-backend")]
         config::CaptureDevice::Pulse {
             channels,
             ref device,
             format,
+            ..
         } => Box::new(pulsedevice::PulseCaptureDevice {
             devname: device.clone(),
             samplerate: conf.samplerate,
@@ -635,14 +640,14 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             stop_on_rate_change: conf.stop_on_rate_change(),
             rate_measure_interval: conf.rate_measure_interval(),
         }),
-        config::CaptureDevice::SignalGenerator { signal, channels } => {
-            Box::new(generatordevice::GeneratorDevice {
-                signal,
-                samplerate: conf.samplerate,
-                channels,
-                chunksize: conf.chunksize,
-            })
-        }
+        config::CaptureDevice::SignalGenerator {
+            signal, channels, ..
+        } => Box::new(generatordevice::GeneratorDevice {
+            signal,
+            samplerate: conf.samplerate,
+            channels,
+            chunksize: conf.chunksize,
+        }),
         #[cfg(all(target_os = "linux", feature = "bluez-backend"))]
         config::CaptureDevice::Bluez(ref dev) => Box::new(filedevice::FileCaptureDevice {
             source: filedevice::CaptureSource::BluezDBus(dev.service(), dev.dbus_path.clone()),
@@ -705,6 +710,7 @@ pub fn new_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
         config::CaptureDevice::Jack {
             channels,
             ref device,
+            ..
         } => Box::new(cpaldevice::CpalCaptureDevice {
             devname: device.clone(),
             host: cpaldevice::CpalHost::Jack,