+++ src/bin.rs
@@ -2,8 +2,6 @@
 extern crate alsa;
 extern crate camillalib;
 extern crate clap;
-#[cfg(feature = "FFTW")]
-extern crate fftw;
 extern crate lazy_static;
 #[cfg(feature = "pulse-backend")]
 extern crate libpulse_binding as pulse;
@@ -12,7 +10,6 @@ extern crate libpulse_simple_binding as psimple;
 extern crate parking_lot;
 extern crate rand;
 extern crate rand_distr;
-#[cfg(not(feature = "FFTW"))]
 extern crate realfft;
 extern crate rubato;
 extern crate serde;
@@ -179,6 +176,7 @@ fn run(
             tx_status_cap,
             rx_command_cap,
             status_structs.capture.clone(),
+            status_structs.processing.clone(),
         )
         .unwrap();
 
@@ -404,6 +402,10 @@ fn run(
                             debug!("SetVolume message to  {} dB received", vol);
                             status_structs.processing.set_target_volume(0, vol);
                         }
+                        StatusMessage::SetMute(mute) => {
+                            debug!("SetMute message to {} received", mute);
+                            status_structs.processing.set_mute(0, mute);
+                        }
                     },
                     Err(err) => {
                         warn!("Capture, Playback and Processing threads have exited: {}", err);
@@ -437,9 +439,6 @@ fn main_process() -> i32 {
     }
     if cfg!(feature = "secure-websocket") {
         features.push("secure-websocket");
-    }
-    if cfg!(feature = "FFTW") {
-        features.push("FFTW");
     }
     if cfg!(feature = "32bit") {
         features.push("32bit");
@@ -515,11 +514,31 @@ fn main_process() -> i32 {
         )
         .arg(
             Arg::new("logfile")
-                .help("Write logs to file")
+                .help("Write logs to the given file path")
                 .short('o')
                 .long("logfile")
                 .value_name("LOGFILE")
-                .display_order(100)
+                .display_order(101)
+                .action(ArgAction::Set),
+        )
+        .arg(
+            Arg::new("log_rotate_size")
+                .help("Rotate log file when the size in bytes exceeds this value")
+                .long("log_rotate_size")
+                .value_name("ROTATE_SIZE")
+                .display_order(102)
+                .requires("logfile")
+                .value_parser(clap::value_parser!(u32).range(1000..))
+                .action(ArgAction::Set),
+        )
+        .arg(
+            Arg::new("log_keep_nbr")
+                .help("Number of previous log files to keep")
+                .long("log_keep_nbr")
+                .value_name("KEEP_NBR")
+                .display_order(103)
+                .requires("log_rotate_size")
+                .value_parser(clap::value_parser!(u32))
                 .action(ArgAction::Set),
         )
         .arg(
@@ -728,13 +747,27 @@ fn main_process() -> i32 {
             fullpath.push(path);
             path = fullpath;
         }
-        flexi_logger::Logger::try_with_str(loglevel)
+        let mut logger = flexi_logger::Logger::try_with_str(loglevel)
             .unwrap()
             .format(custom_logger_format)
             .log_to_file(flexi_logger::FileSpec::try_from(path).unwrap())
-            .write_mode(flexi_logger::WriteMode::Async)
-            .start()
-            .unwrap()
+            .write_mode(flexi_logger::WriteMode::Async);
+
+        let cleanup = if let Some(keep_nbr) = matches.get_one::<u32>("log_keep_nbr") {
+            flexi_logger::Cleanup::KeepLogFiles(*keep_nbr as usize)
+        } else {
+            flexi_logger::Cleanup::Never
+        };
+
+        if let Some(rotate_size) = matches.get_one::<u32>("log_rotate_size") {
+            logger = logger.rotate(
+                flexi_logger::Criterion::Size(*rotate_size as u64),
+                flexi_logger::Naming::Timestamps,
+                cleanup,
+            );
+        }
+
+        logger.start().unwrap()
     } else {
         flexi_logger::Logger::try_with_str(loglevel)
             .unwrap()