+++ src/config.rs
@@ -198,7 +198,11 @@ pub enum CaptureDevice {
         #[serde(default)]
         stop_on_inactive: Option<bool>,
         #[serde(default)]
-        follow_volume_control: Option<String>,
+        link_volume_control: Option<String>,
+        #[serde(default)]
+        link_mute_control: Option<String>,
+        #[serde(default)]
+        labels: Option<Vec<Option<String>>>,
     },
     #[cfg(all(target_os = "linux", feature = "bluez-backend"))]
     #[serde(alias = "BLUEZ", alias = "bluez")]
@@ -210,6 +214,8 @@ pub enum CaptureDevice {
         channels: usize,
         device: String,
         format: SampleFormat,
+        #[serde(default)]
+        labels: Option<Vec<Option<String>>>,
     },
     RawFile(CaptureDeviceRawFile),
     WavFile(CaptureDeviceWavFile),
@@ -236,11 +242,15 @@ pub enum CaptureDevice {
         #[serde(deserialize_with = "validate_nonzero_usize")]
         channels: usize,
         device: String,
+        #[serde(default)]
+        labels: Option<Vec<Option<String>>>,
     },
     SignalGenerator {
         #[serde(deserialize_with = "validate_nonzero_usize")]
         channels: usize,
         signal: Signal,
+        #[serde(default)]
+        labels: Option<Vec<Option<String>>>,
     },
 }
 
@@ -291,6 +301,8 @@ pub struct CaptureDeviceRawFile {
     pub skip_bytes: Option<usize>,
     #[serde(default)]
     pub read_bytes: Option<usize>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 impl CaptureDeviceRawFile {
@@ -311,6 +323,8 @@ pub struct CaptureDeviceWavFile {
     pub filename: String,
     #[serde(default)]
     pub extra_samples: Option<usize>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 impl CaptureDeviceWavFile {
@@ -344,6 +358,8 @@ pub struct CaptureDeviceStdin {
     pub skip_bytes: Option<usize>,
     #[serde(default)]
     pub read_bytes: Option<usize>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 impl CaptureDeviceStdin {
@@ -370,6 +386,8 @@ pub struct CaptureDeviceBluez {
     // from D-Bus properties
     pub format: SampleFormat,
     pub channels: usize,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 #[cfg(all(target_os = "linux", feature = "bluez-backend"))]
@@ -391,6 +409,8 @@ pub struct CaptureDeviceWasapi {
     exclusive: Option<bool>,
     #[serde(default)]
     loopback: Option<bool>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 #[cfg(target_os = "windows")]
@@ -413,6 +433,8 @@ pub struct CaptureDeviceCA {
     pub device: Option<String>,
     #[serde(default)]
     pub format: Option<SampleFormat>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
@@ -575,6 +597,10 @@ pub struct Devices {
     pub volume_ramp_time: Option<f32>,
     #[serde(default)]
     pub volume_limit: Option<f32>,
+    #[serde(default)]
+    pub multithreaded: Option<bool>,
+    #[serde(default)]
+    pub worker_threads: Option<usize>,
 }
 
 // Getters for all the defaults
@@ -621,6 +647,14 @@ impl Devices {
 
     pub fn volume_limit(&self) -> f32 {
         self.volume_limit.unwrap_or(50.0)
+    }
+
+    pub fn multithreaded(&self) -> bool {
+        self.multithreaded.unwrap_or(false)
+    }
+
+    pub fn worker_threads(&self) -> usize {
+        self.worker_threads.unwrap_or(0)
     }
 }
 
@@ -1208,6 +1242,8 @@ pub struct Mixer {
     pub description: Option<String>,
     pub channels: MixerChannels,
     pub mapping: Vec<MixerMapping>,
+    #[serde(default)]
+    pub labels: Option<Vec<Option<String>>>,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
@@ -1342,6 +1378,7 @@ pub enum PipelineStep {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
 pub struct PipelineStepMixer {
     pub name: String,
     #[serde(default)]
@@ -1357,6 +1394,7 @@ impl PipelineStepMixer {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
 pub struct PipelineStepFilter {
     #[serde(default)]
     pub channels: Option<Vec<usize>>,
@@ -1374,6 +1412,7 @@ impl PipelineStepFilter {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
 pub struct PipelineStepProcessor {
     pub name: String,
     #[serde(default)]
@@ -1823,8 +1862,8 @@ pub fn validate_config(conf: &mut Configuration, filename: Option<&str>) -> Res<
     #[cfg(not(target_os = "linux"))]
     let target_level_limit = 2 * conf.devices.chunksize;
 
-    if conf.devices.target_level() >= target_level_limit {
-        let msg = format!("target_level can't be larger than {}", target_level_limit);
+    if conf.devices.target_level() > target_level_limit {
+        let msg = format!("target_level cannot be larger than {}", target_level_limit);
         return Err(ConfigError::new(&msg).into());
     }
     if let Some(period) = conf.devices.adjust_period {