+++ src/filters.rs
@@ -7,10 +7,7 @@ use crate::config;
 use crate::conversions;
 use crate::diffeq;
 use crate::dither;
-#[cfg(not(feature = "FFTW"))]
 use crate::fftconv;
-#[cfg(feature = "FFTW")]
-use crate::fftconv_fftw as fftconv;
 use crate::limiter;
 use crate::loudness;
 use crate::mixer;
@@ -23,6 +20,8 @@ use std::io::BufReader;
 use std::io::{BufRead, Seek, SeekFrom};
 use std::sync::Arc;
 use std::time::Instant;
+
+use rayon::prelude::*;
 
 use crate::PrcFmt;
 use crate::ProcessingParameters;
@@ -174,7 +173,7 @@ pub fn read_wav(filename: &str, channel: usize) -> Res<Vec<PrcFmt>> {
 
 pub struct FilterGroup {
     channel: usize,
-    filters: Vec<Box<dyn Filter>>,
+    filters: Vec<Box<dyn Filter + Send>>,
 }
 
 impl FilterGroup {
@@ -188,11 +187,11 @@ impl FilterGroup {
         processing_params: Arc<ProcessingParameters>,
     ) -> Self {
         debug!("Build filter group from config");
-        let mut filters = Vec::<Box<dyn Filter>>::new();
+        let mut filters = Vec::<Box<dyn Filter + Send>>::new();
         for name in names {
             let filter_cfg = filter_configs[name].clone();
             trace!("Create filter {} with config {:?}", name, filter_cfg);
-            let filter: Box<dyn Filter> =
+            let filter: Box<dyn Filter + Send> =
                 match filter_cfg {
                     config::Filter::Conv { parameters, .. } => Box::new(
                         fftconv::FftConv::from_config(name, waveform_length, parameters),
@@ -290,11 +289,46 @@ impl FilterGroup {
     }
 }
 
+pub struct ParallelFilters {
+    filters: Vec<Vec<Box<dyn Filter + Send>>>,
+}
+
+impl ParallelFilters {
+    pub fn update_parameters(
+        &mut self,
+        filterconfigs: HashMap<String, config::Filter>,
+        changed: &[String],
+    ) {
+        for channel_filters in &mut self.filters {
+            for filter in channel_filters {
+                if changed.iter().any(|n| n == filter.name()) {
+                    filter.update_parameters(filterconfigs[filter.name()].clone());
+                }
+            }
+        }
+    }
+
+    /// Apply all the filters to an AudioChunk.
+    fn process_chunk(&mut self, input: &mut AudioChunk) -> Res<()> {
+        self.filters
+            .par_iter_mut()
+            .zip(input.waveforms.par_iter_mut())
+            .filter(|(f, w)| !f.is_empty() && !w.is_empty())
+            .for_each(|(f, w)| {
+                for filt in f {
+                    let _ = filt.process_waveform(w);
+                }
+            });
+        Ok(())
+    }
+}
+
 /// A Pipeline is made up of a series of PipelineSteps,
-/// each one can be a single Mixer of a group of Filters
+/// each one can be a single Mixer or a group of Filters
 pub enum PipelineStep {
     MixerStep(mixer::Mixer),
     FilterStep(FilterGroup),
+    ParallelFiltersStep(ParallelFilters),
     ProcessorStep(Box<dyn Processor>),
 }
 
@@ -410,6 +444,9 @@ impl Pipeline {
             0,
         );
         let secs_per_chunk = conf.devices.chunksize as f32 / conf.devices.samplerate as f32;
+        if conf.devices.multithreaded() {
+            steps = parallelize_filters(&mut steps, conf.devices.capture.channels());
+        }
         Pipeline {
             steps,
             volume,
@@ -436,6 +473,9 @@ impl Pipeline {
                 PipelineStep::FilterStep(flt) => {
                     flt.update_parameters(conf.filters.as_ref().unwrap().clone(), filters);
                 }
+                PipelineStep::ParallelFiltersStep(flt) => {
+                    flt.update_parameters(conf.filters.as_ref().unwrap().clone(), filters);
+                }
                 PipelineStep::ProcessorStep(proc) => {
                     if processors.iter().any(|n| n == proc.name()) {
                         proc.update_parameters(
@@ -459,6 +499,9 @@ impl Pipeline {
                 PipelineStep::FilterStep(flt) => {
                     flt.process_chunk(&mut chunk).unwrap();
                 }
+                PipelineStep::ParallelFiltersStep(flt) => {
+                    flt.process_chunk(&mut chunk).unwrap();
+                }
                 PipelineStep::ProcessorStep(comp) => {
                     comp.process_chunk(&mut chunk).unwrap();
                 }
@@ -470,6 +513,67 @@ impl Pipeline {
         trace!("Processing load: {load}%");
         chunk
     }
+}
+
+// Loop trough the pipeline to merge individual filter steps,
+// in order use rayon to apply them in parallel.
+fn parallelize_filters(steps: &mut Vec<PipelineStep>, nbr_channels: usize) -> Vec<PipelineStep> {
+    debug!("Merging filter steps to enable parallel processing");
+    let mut new_steps: Vec<PipelineStep> = Vec::new();
+    let mut parfilt = None;
+    let mut active_channels = nbr_channels;
+    for step in steps.drain(..) {
+        match step {
+            PipelineStep::MixerStep(ref mix) => {
+                if parfilt.is_some() {
+                    debug!("Append parallel filter step to pipeline");
+                    new_steps.push(PipelineStep::ParallelFiltersStep(parfilt.take().unwrap()));
+                }
+                active_channels = mix.channels_out;
+                debug!("Append mixer step to pipeline");
+                new_steps.push(step);
+            }
+            PipelineStep::ProcessorStep(_) => {
+                if parfilt.is_some() {
+                    debug!("Append parallel filter step to pipeline");
+                    new_steps.push(PipelineStep::ParallelFiltersStep(parfilt.take().unwrap()));
+                }
+                debug!("Append processor step to pipeline");
+                new_steps.push(step);
+            }
+            PipelineStep::ParallelFiltersStep(_) => {
+                if parfilt.is_some() {
+                    debug!("Append parallel filter step to pipeline");
+                    new_steps.push(PipelineStep::ParallelFiltersStep(parfilt.take().unwrap()));
+                }
+                debug!("Append existing parallel filter step to pipeline");
+                new_steps.push(step);
+            }
+            PipelineStep::FilterStep(mut flt) => {
+                if parfilt.is_none() {
+                    debug!("Start new parallel filter step");
+                    let mut filters = Vec::with_capacity(active_channels);
+                    for _ in 0..active_channels {
+                        filters.push(Vec::new());
+                    }
+                    parfilt = Some(ParallelFilters { filters });
+                }
+                if let Some(ref mut f) = parfilt {
+                    debug!(
+                        "Adding {} filters to channel {} of parallel filter step",
+                        flt.filters.len(),
+                        flt.channel
+                    );
+                    f.filters[flt.channel].append(&mut flt.filters);
+                }
+            }
+        }
+    }
+    if parfilt.is_some() {
+        debug!("Append parallel filter step to pipeline");
+        new_steps.push(PipelineStep::ParallelFiltersStep(parfilt.take().unwrap()));
+    }
+    new_steps
 }
 
 /// Validate the filter config, to give a helpful message intead of a panic.